app/hello_django/serializers.py
from rest_framework import serializers

from app.hello_django.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'name', 'surname', 'email')
