app/requirements.txt
Django==4.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
djangorestframework==3.14
